Added extern macro for importing the configuration.
